+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+Release 5.1.0:
+ - Update JsonPath4K to 2.4.0
+ - Fix XCF export with transitive dependencies
+
+Release 5.0.1:
+ - Fix verifiable presentation of ISO credentials to contain `DeviceResponse` instead of a `Document`
+ - Data classes for verification result of ISO structures now may contain more than one document
+
 Release 5.0.0:
  - Remove `OidcSiopWallet.newDefaultInstance()` and replace it with a constructor
  - Remove `OidcSiopVerifier.newInstance()` methods and replace them with constructors